2025-01 이후의 작업 77개가 jpajoin 에만 쌓여 있어 master 가 뒤처져 있었다.
master 쪽에 고유한 변경은 없어(트리가 merge-base 와 동일) 충돌 없이 합쳐진다.

앞으로는 master 에서 작업한다.

+++ reactboot/src/main/java/com/example/firstproject/Dto/Comment/Commentform.java
@@ -22,7 +22,7 @@ public class Commentform {
 	@NotBlank
 	private int depth;
 	
-	private int cnum;
+	private Long cnum;
 	
 
 	@NotBlank
